# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vox Trader Application API",
    "description": "The application boundary between the Vox backend and its clients. The browser talks to this API and never to a broker. Money crosses as exact decimal strings, never as JSON numbers. A capability without a backend owner answers CAPABILITY_UNAVAILABLE rather than a plausible success.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment"
    }
  ],
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Accounts discovered through the connection.",
        "operationId": "accounts",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Provider of the connection, for example `T_INVEST`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderDto"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "description": "Broker environment: `SANDBOX` or `PRODUCTION`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BrokerEnvironment"
            }
          },
          {
            "name": "broker_account_id",
            "in": "query",
            "description": "Broker account identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connection_ref",
            "in": "query",
            "description": "Opaque connection reference. Never a credential.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrokerAccountDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "No account read side is attached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capabilities": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "What this deployment can actually do.",
        "operationId": "capabilities",
        "parameters": [
          {
            "name": "broker_account_id",
            "in": "query",
            "description": "Scope the capability set to one account",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Capability set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitySet"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commands/cancel-order": {
      "post": {
        "tags": [
          "execution"
        ],
        "summary": "Cancel a regular order.",
        "operationId": "cancel_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationReceiptDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commands/order": {
      "post": {
        "tags": [
          "execution"
        ],
        "summary": "Submit a regular order. The scope in the body is the frozen target of the command.",
        "operationId": "submit_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Mutation receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationReceiptDto"
                }
              }
            }
          },
          "202": {
            "description": "Dispatched, outcome not yet known",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Execution is not authorized for this scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "No execution side is attached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Operations history, paged by cursor.",
        "operationId": "operations",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderDto"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BrokerEnvironment"
            }
          },
          {
            "name": "broker_account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connection_ref",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Cursor returned by the previous page.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size; the server caps it.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsPageDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portfolio": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Currency balances as the broker reports them.",
        "operationId": "portfolio",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Provider of the connection, for example `T_INVEST`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderDto"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "description": "Broker environment: `SANDBOX` or `PRODUCTION`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BrokerEnvironment"
            }
          },
          {
            "name": "broker_account_id",
            "in": "query",
            "description": "Broker account identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connection_ref",
            "in": "query",
            "description": "Opaque connection reference. Never a credential.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/positions": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Positions and their quantities.",
        "operationId": "positions",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Provider of the connection, for example `T_INVEST`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderDto"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "description": "Broker environment: `SANDBOX` or `PRODUCTION`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BrokerEnvironment"
            }
          },
          {
            "name": "broker_account_id",
            "in": "query",
            "description": "Broker account identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connection_ref",
            "in": "query",
            "description": "Opaque connection reference. Never a credential.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconciliation": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "How complete the last reconciliation was.",
        "operationId": "reconciliation",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Provider of the connection, for example `T_INVEST`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderDto"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "description": "Broker environment: `SANDBOX` or `PRODUCTION`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BrokerEnvironment"
            }
          },
          {
            "name": "broker_account_id",
            "in": "query",
            "description": "Broker account identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connection_ref",
            "in": "query",
            "description": "Opaque connection reference. Never a credential.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runtime": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Runtime state, readiness and stream health.",
        "operationId": "runtime_health",
        "responses": {
          "200": {
            "description": "Current runtime health",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeHealthDto"
                }
              }
            }
          },
          "503": {
            "description": "No runtime is attached to this process",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stop-orders": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Stop orders.",
        "operationId": "stop_orders",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Provider of the connection, for example `T_INVEST`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderDto"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "description": "Broker environment: `SANDBOX` or `PRODUCTION`.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BrokerEnvironment"
            }
          },
          {
            "name": "broker_account_id",
            "in": "query",
            "description": "Broker account identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connection_ref",
            "in": "query",
            "description": "Opaque connection reference. Never a credential.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StopOrderDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "The public error envelope. Never carries provider payloads, credentials or stack traces.",
        "required": [
          "code",
          "message",
          "correlation_id",
          "category",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine code, screaming snake case.",
            "example": "CAPABILITY_UNAVAILABLE"
          },
          "message": {
            "type": "string",
            "description": "Human sentence for the operator."
          },
          "correlation_id": {
            "type": "string",
            "description": "Correlates this response with server logs and with a mutation, when there is one."
          },
          "category": {
            "$ref": "#/components/schemas/ErrorCategory"
          },
          "retryable": {
            "type": "boolean"
          },
          "field_errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "details": {
            "description": "Safe, typed extra context. Never provider metadata."
          }
        }
      },
      "BrokerAccountDto": {
        "type": "object",
        "description": "A broker account discovered through a connection.",
        "required": [
          "account_id",
          "open",
          "accessible"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "open": {
            "type": "boolean"
          },
          "accessible": {
            "type": "boolean",
            "description": "Whether the runtime can currently read this account."
          }
        }
      },
      "BrokerEnvironment": {
        "type": "string",
        "description": "The broker-side environment of a connection. Exactly the runtime contract's two values.",
        "enum": [
          "SANDBOX",
          "PRODUCTION"
        ]
      },
      "CancelOrderRequest": {
        "type": "object",
        "description": "Cancel a regular order by broker identity or by the logical request that created it.",
        "required": [
          "scope",
          "client_request_id"
        ],
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/ExecutionScope"
          },
          "client_request_id": {
            "type": "string"
          },
          "broker_order_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "logical_request_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CapabilitySet": {
        "type": "object",
        "description": "What this deployment can actually do, for one scope.",
        "required": [
          "provider",
          "environment",
          "supported",
          "unavailable"
        ],
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderDto"
          },
          "environment": {
            "$ref": "#/components/schemas/BrokerEnvironment"
          },
          "account_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Present when the capability set is account-scoped."
          },
          "supported": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Capability"
            }
          },
          "unavailable": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnavailableCapability"
            }
          }
        }
      },
...
      "MutationReceiptDto": {
        "type": "object",
        "description": "The receipt of a capital-affecting command.",
        "required": [
          "logical_request_id",
          "scope",
          "kind",
          "state",
          "decision",
          "correlation_id",
          "runtime_epoch",
          "created_at_unix_ms",
          "updated_at_unix_ms"
        ],
        "properties": {
          "logical_request_id": {
            "type": "string",
            "description": "Vox-side identity of the command, stable across retries and reconciliation."
          },
          "scope": {
            "$ref": "#/components/schemas/ExecutionScope",
            "description": "The frozen target this command was dispatched to."
          },
          "kind": {
            "$ref": "#/components/schemas/MutationKindDto"
          },
          "state": {
            "$ref": "#/components/schemas/JournalStateDto"
          },
          "decision": {
            "$ref": "#/components/schemas/MutationDecisionDto",
            "description": "What the backend says the client may do next."
          },
          "correlation_id": {
            "type": "string"
          },
          "broker_order_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set only once the broker has answered."
          },
          "broker_stop_order_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "reconciliation_disposition": {
            "type": [
              "string",
              "null"
            ],
            "description": "Runtime disposition after reconciliation, when it has run."
          },
          "runtime_epoch": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "created_at_unix_ms": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at_unix_ms": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "OperationsPageDto": {
        "type": "object",
        "description": "A page of operations with the cursor to continue from.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OperationDto"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PortfolioDto": {
        "type": "object",
        "description": "Portfolio as the broker reports it: currency balances and when they were observed.\n\nValuation, P&L, exposure and margin are **not** here. #22 owns them.",
        "required": [
          "account_id",
          "balances"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "balances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurrencyBalanceDto"
            }
          },
          "broker_observed_at_unix_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "PositionDto": {
        "type": "object",
        "description": "A position as the broker reports it: instrument and quantity, nothing derived.",
        "required": [
          "account_id",
          "instrument_uid",
          "quantity_units"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "instrument_uid": {
            "type": "string"
          },
          "quantity_units": {
            "type": "integer",
            "format": "int64",
            "description": "Signed quantity in instrument units."
          },
          "broker_observed_at_unix_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "ProviderDto": {
        "type": "string",
        "description": "Providers with a registered adapter. A provider appears here only once it is real.",
        "enum": [
          "T_INVEST"
        ]
      },
...
      "ReconciliationDto": {
        "type": "object",
        "description": "How complete the last reconciliation was, per domain.",
        "required": [
          "scope_key",
          "reconciliation_id",
          "snapshot_observed_at_unix_ms",
          "completed_at_unix_ms",
          "runtime_epoch",
          "accounts_complete",
          "portfolio_complete",
          "positions_complete",
          "orders_complete",
          "stops_complete",
          "operations_complete",
          "complete"
        ],
        "properties": {
          "scope_key": {
            "type": "string"
          },
          "reconciliation_id": {
            "type": "string"
          },
          "operations_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "snapshot_observed_at_unix_ms": {
            "type": "integer",
            "format": "int64"
          },
          "completed_at_unix_ms": {
            "type": "integer",
            "format": "int64"
          },
          "runtime_epoch": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "accounts_complete": {
            "type": "boolean"
          },
          "portfolio_complete": {
            "type": "boolean"
          },
          "positions_complete": {
            "type": "boolean"
          },
          "orders_complete": {
            "type": "boolean"
          },
          "stops_complete": {
            "type": "boolean"
          },
          "operations_complete": {
            "type": "boolean"
          },
          "complete": {
            "type": "boolean",
            "description": "True only when every domain above is complete."
          }
        }
      },
...
      "StopOrderDto": {
        "type": "object",
        "description": "A stop order identity and its liveness. Trigger levels are not in the read model yet.",
        "required": [
          "account_id",
          "broker_stop_order_id",
          "instrument_uid",
          "active",
          "terminal"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "broker_stop_order_id": {
            "type": "string"
          },
          "logical_request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "instrument_uid": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "terminal": {
            "type": "boolean"
          }
        }
      },
...
      "SubmitOrderRequest": {
        "type": "object",
        "description": "Submit a regular order. Quantity is in lots; price is exact and optional for market orders.",
        "required": [
          "scope",
          "instrument_uid",
          "client_request_id",
          "side",
          "order_type",
          "quantity_lots",
          "price_convention",
          "time_in_force"
        ],
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/ExecutionScope",
            "description": "The immutable target. A submitted command is never retargeted by a later UI change."
          },
          "instrument_uid": {
            "type": "string"
          },
          "client_request_id": {
            "type": "string",
            "description": "Client-generated identity of this command, used for idempotency and reconciliation.",
            "example": "b4f1c2a0-6d18-4f0e-9a37-2d5c1f0b7e44"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSideDto"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderTypeDto"
          },
          "quantity_lots": {
            "type": "integer",
            "format": "int64"
          },
          "price": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Decimal"
              }
            ]
          },
          "price_convention": {
            "$ref": "#/components/schemas/PriceConventionDto"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForceDto"
          },
          "protection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProtectionPlanDto"
              }
            ]
          }
        }
      },
...
    }
  },
...
}